+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of lizard-nxt client
 Unreleased (1.5.0) (XXXX-XX-XX)
 -------------------------------
 
-- Add search to time.
+- Add date parsing to search bar.
 
 
 Release 1.4.1 (2015-5-1)